refactor: moves all "/exports.ts" with functional "tool" to "/exports_toolbox.ts"
