+++ src/library/math/operator/constructive/absoluteValueOf/aliases/index.ts
@@ -1,1 +1,1 @@
-export * from './exports.ts';
+export * from './exports_toolbox.ts';
