+++ src/library/math/operator/constructive/greatestCommonDivisor/abbreviations/exports_toolbox.ts
