+++ src/library/math/operator/constructive/greatestCommonDivisor/aliases/exports_toolbox.ts
