+++ src/library/math/operator/constructive/greatestCommonDivisor/aliases/index.ts
@@ -1,1 +1,1 @@
-export * from './exports.ts';
+export * from './exports_toolbox.ts';
