+++ src/library/math/operator/hyper/rank3/index.ts
@@ -1,1 +1,1 @@
-export * from './exports.ts';
+export * from './exports_toolbox.ts';
