+++ src/library/math/operator/predicate/isWhole/aliases/exports_toolbox.ts
