+++ src/library/typeUtilities/index.ts
@@ -1,1 +1,1 @@
-export type * from './exports.ts';
+export type * from './exports_toolbox.ts';
