+++ src/library/utilitiesByType/string/empty/index.ts
@@ -1,1 +1,1 @@
-export * from './exports.ts';
+export * from './exports_toolbox.ts';
